chore(tests): move test/local/lib/* up to test/local/
