+++ test/local/senders/reminder.js
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const ROOT_DIR = '../../../../'
+const ROOT_DIR = '../../../'
 
 const assert = require('insist')
 var sinon = require('sinon')
